+++ source/labyrinths/zooofchimeras.js
@@ -193,7 +193,7 @@ module.exports = new LabyrinthTemplate("Zoo of Chimeras",
 				"Accelerating Refreshing Breeze",
 				"Supportive Refreshing Breeze",
 				"Swift Refreshing Breeze",
-				"Accelerating Air Blades",
+				"Adventurer's Air Blades",
 				"Toxic Air Blades",
 				"Unstoppable Air Blades"
 			]